Create custom unmarshaler
Custom unmarshaler was necessary since steam spy's 'tags' value
was an array of key-val pairs. If that array was empty, it had
trouble unmarshaling an empty array into a map.
Created a custom unmarshaler to dismiss that issue by
assigning a nil map (instead of an empty array) to tags map.